Global reshape : amc2moodle folder.
Move amc2moodle and moodle2amc into pkg amc2moodle folder.

Reshape. Fix many path problems.

Update github action files with moodle2amc and new paths.
