+++ MANIFEST.in
@@ -1,14 +1,18 @@
-include amc2moodle/bin/amc2moodle
-include amc2moodle/*.xslt
-include amc2moodle/*sty.ltxml
-include amc2moodle/test/Figures/*
-include amc2moodle/test/Figures/other/*
-include amc2moodle/test/QCM.pdf
-include amc2moodle/test/QCM.tex
-include amc2moodle/test/QCM.xml
-include amc2moodle/test/QCM_wo-tikz.tex
-include amc2moodle/test/QCM_wo-tikz.xml
 include LICENSE
 include README.md
+include amc2moodle/amc2moodle/bin/amc2moodle
+include amc2moodle/amc2moodle/*.xslt
+include amc2moodle/amc2moodle/*sty.ltxml
+include amc2moodle/amc2moodle/test/Figures/*
+include amc2moodle/amc2moodle/test/Figures/other/*
+include amc2moodle/amc2moodle/test/QCM.pdf
+include amc2moodle/amc2moodle/test/QCM.tex
+include amc2moodle/amc2moodle/test/QCM.xml
+include amc2moodle/amc2moodle/test/QCM_wo-tikz.tex
+include amc2moodle/amc2moodle/test/QCM_wo-tikz.xml
+include amc2moodle/moodle2amc/bin/moodle2amc
+include amc2moodle/moodle2amc/html2tex.xslt
+include amc2moodle/moodle2amc/struc2tex.xslt
+
