+++ amc2moodle/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '2.0.0b1'
+__version__ = '2.0.0b2'
