+++ amc2moodle/amc2moodle/amc2moodle_class.py
@@ -18,7 +18,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from amc2moodle import convert
+from amc2moodle.amc2moodle import convert
 import subprocess
 import sys
 import os
