+++ amc2moodle/amc2moodle/automultiplechoice.sty.ltxml
