+++ amc2moodle/amc2moodle/test/Figures/tinymonk.pdf
