+++ amc2moodle/moodle2amc/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Thu May 21 15:15:44 2020
+
+@author: bn
+"""
+
+# import class
+from ._questions import *
+from ._quiz import Quiz
