+++ setup.py
@@ -50,9 +50,10 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/nennigb/amc2moodle",
     zip_safe=False,
-    packages=['amc2moodle'],
+    packages=['amc2moodle.amc2moodle', 'amc2moodle.moodle2amc'],
     include_package_data=True,
-    scripts=['amc2moodle/bin/amc2moodle'],
+    scripts=['amc2moodle/amc2moodle/bin/amc2moodle',
+             'amc2moodle/moodle2amc/bin/moodle2amc'],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: GNU GPL 3",
